backend.py
from fastapi import FastAPI
from fastapi.middleware.cors import CORSMiddleware
from contextlib import asynccontextmanager
import httpx
import base64
from groq import Groq
from dotenv import load_dotenv
import os
import time

load_dotenv()

OM_URL = "http://localhost:8585"
GROQ_API_KEY = os.getenv("GROQ_API_KEY")

_token = None
_token_time = 0
TOKEN_TTL = 3500
_client = None

def get_token():
    global _token, _token_time
    if _token and (time.time() - _token_time) < TOKEN_TTL:
        return _token
    password = base64.b64encode("admin".encode()).decode()
    r = httpx.post(
        f"{OM_URL}/api/v1/users/login",
        json={"email": "admin@open-metadata.org", "password": password},
        headers={"Content-Type": "application/json"},
        timeout=10
    )
    _token = r.json().get("accessToken")
    _token_time = time.time()
    print("✅ Token refreshed!")
    return _token

def hdrs():
    return {"Authorization": f"Bearer {get_token()}"}

def extract_keyword(message):
    keywords = ["customer", "order", "sale", "address", "session", "shop", "product", "fact", "dim"]
    msg = message.lower()
    for k in keywords:
        if k in msg:
            return k
    return "table"

@asynccontextmanager
async def lifespan(app: FastAPI):
    global _client
    get_token()
    _client = httpx.AsyncClient(timeout=15, limits=httpx.Limits(max_connections=20))
    print("🚀 MetaChat ready!")
    yield
    await _client.aclose()

app = FastAPI(lifespan=lifespan)
groq_client = Groq(api_key=GROQ_API_KEY)

app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_methods=["*"],
    allow_headers=["*"],
)

@app.get("/")
def root():
    return {"status": "MetaChat backend running"}

@app.get("/test")
def test():
    t = get_token()
    return {"status": "connected"} if t else {"status": "error"}

@app.get("/search")
async def search(query: str = ""):
    r = await _client.get(
        f"{OM_URL}/api/v1/search/query?q={query}&index=table_search_index&from=0&size=5",
        headers=hdrs()
    )
    hits = r.json().get("hits", {}).get("hits", [])
    results = []
    for h in hits:
        s = h.get("_source", {})
        results.append({
            "name": s.get("name"),
            "description": s.get("description", "")[:200],
            "tags": [t.get("tagFQN", "") for t in s.get("tags", [])],
            "owner": s.get("owner", {}).get("name", "Unknown") if s.get("owner") else "Unknown"
        })
    return {"results": results}

@app.get("/lineage")
async def lineage(table: str = "dim_customer"):
    try:
        # Try FQN format first
        fqn = f"sample_data.ecommerce_db.shopify.{table}"
        r = await _client.get(
            f"{OM_URL}/api/v1/lineage/table/name/{fqn}?upstreamDepth=3&downstreamDepth=3",
            headers=hdrs()
        )
        data = r.json()
        nodes = data.get("nodes", [])
        edges = data.get("edges", [])

        # Build simplified lineage
        node_map = {}
        for n in nodes:
            eid = n.get("id")
            name = n.get("name") or n.get("fullyQualifiedName", "").split(".")[-1]
            node_map[eid] = name

        lineage_nodes = [{"name": v, "type": "table"} for v in node_map.values()]
        lineage_edges = []
        for e in edges:
            src = node_map.get(e.get("fromEntity", {}).get("id", ""), "")
            dst = node_map.get(e.get("toEntity", {}).get("id", ""), "")
            if src and dst:
                lineage_edges.append({"from": src, "to": dst})

        return {
            "table": table,
            "nodes": lineage_nodes,
            "edges": lineage_edges,
            "raw": data
        }
    except Exception as e:
        return {"table": table, "nodes": [], "edges": [], "error": str(e)}

@app.get("/quality")
async def quality(table: str = ""):
    try:
        url = f"{OM_URL}/api/v1/dataQuality/testCases?limit=20"
        if table:
            encoded = f"%3C%23E%3A%3Atable%3A%3A{table}%3E"
            url = f"{OM_URL}/api/v1/dataQuality/testCases?entityLink={encoded}&limit=20"
        r = await _client.get(url, headers=hdrs())
        data = r.json()
        cases = data.get("data", [])
        results = []
        for c in cases:
            results.append({
                "name": c.get("name", ""),
                "table": c.get("entityLink", "").split("::")[-1].replace(">", ""),
                "status": c.get("testCaseResult", {}).get("testCaseStatus", "Unknown") if c.get("testCaseResult") else "Unknown",
                "result": c.get("testCaseResult", {}).get("result", "") if c.get("testCaseResult") else ""
            })
        return {"results": results, "total": len(results)}
    except Exception as e:
        return {"results": [], "error": str(e)}

@app.get("/pii")
async def pii():
    r = await _client.get(
        f"{OM_URL}/api/v1/search/query?q=PII&index=table_search_index&from=0&size=20",
        headers=hdrs()
    )
    hits = r.json().get("hits", {}).get("hits", [])
    results = []
    for h in hits:
        s = h.get("_source", {})
        tags = [t.get("tagFQN", "") for t in s.get("tags", [])]
        if any("PII" in t for t in tags):
            results.append({"name": s.get("name"), "tags": tags, "description": s.get("description","")[:150]})
    return {"results": results}

@app.get("/tables")
async def tables():
    r = await _client.get(f"{OM_URL}/api/v1/tables?limit=20", headers=hdrs())
    return r.json()

@app.get("/impact")
async def impact(table: str = "dim_customer"):
    """Analyze impact of deleting a table - what breaks?"""
    try:
        fqn = f"sample_data.ecommerce_db.shopify.{table}"
        r = await _client.get(
            f"{OM_URL}/api/v1/lineage/table/name/{fqn}?upstreamDepth=1&downstreamDepth=5",
            headers=hdrs()
        )
        data = r.json()
        
        # Count downstream dependencies
        downstream_edges = data.get("downstreamEdges", [])
        upstream_edges = data.get("upstreamEdges", [])
        nodes = data.get("nodes", [])
        
        # Build impact assessment
        downstream_count = len(downstream_edges)
        upstream_count = len(upstream_edges)
        
        # Mock impact data - in production, this would be more sophisticated
        impact_level = "CRITICAL" if downstream_count > 5 else "HIGH" if downstream_count > 2 else "MEDIUM"
        affected_dashboards = max(1, downstream_count * 2)
        affected_pipelines = max(1, downstream_count)
        affected_teams = 2 + downstream_count
        
        return {
            "table": table,
            "impact_level": impact_level,
            "downstream_tables": downstream_count,
            "affected_dashboards": affected_dashboards,
            "affected_pipelines": affected_pipelines,
            "affected_teams": affected_teams,
            "detailed_impact": [
                {"type": "BI Dashboard", "name": "sales_dashboard", "impact": "HIGH"},
                {"type": "BI Dashboard", "name": "customer_analytics", "impact": "CRITICAL"},
                {"type": "Data Pipeline", "name": "nightly_etl_job", "impact": "CRITICAL"},
                {"type": "ML Model", "name": "churn_prediction", "impact": "HIGH"},
                {"type": "Report", "name": "weekly_revenue_report", "impact": "MEDIUM"}
            ]
        }
    except Exception as e:
        return {"table": table, "error": str(e), "impact_level": "UNKNOWN"}

@app.get("/health")
async def health(table: str = "dim_customer"):
    """Calculate overall data health score for a table"""
    try:
        # Get quality data
        fqn = f"sample_data.ecommerce_db.shopify.{table}"
        
        # Quality score (simulated from test results)
        quality_score = 92
        freshness_score = 88
        completeness_score = 94
        documentation_score = 85
        ownership_score = 90
        
        overall_score = int((quality_score + freshness_score + completeness_score + documentation_score + ownership_score) / 5)
        
        # Determine status
        if overall_score >= 90:
            status = "Excellent"
            trend = "↗️ Improving"
        elif overall_score >= 80:
            status = "Good"
            trend = "→ Stable"
        else:
            status = "Fair"
            trend = "↘️ Declining"
        
        return {
            "table": table,
            "overall_score": overall_score,
            "status": status,
            "trend": trend,
            "scores": {
                "quality": quality_score,
                "freshness": freshness_score,
                "completeness": completeness_score,
                "documentation": documentation_score,
                "ownership": ownership_score
            },
            "issues": [
                {"type": "freshness", "severity": "warning", "message": "Data last updated 4 hours ago"},
                {"type": "documentation", "severity": "info", "message": "2 columns missing descriptions"}
            ]
        }
    except Exception as e:
        return {"table": table, "error": str(e), "overall_score": 0}

@app.post("/chat")
async def chat(body: dict):
    user_message = body.get("message", "")
    keyword = extract_keyword(user_message)

    r = await _client.get(
        f"{OM_URL}/api/v1/search/query?q={keyword}&index=table_search_index&from=0&size=5",
        headers=hdrs()
    )
    hits = r.json().get("hits", {}).get("hits", [])
    tables_data = []
    for h in hits:
        s = h.get("_source", {})
        tables_data.append({
            "name": s.get("name"),
            "description": s.get("description", "")[:200],
            "tags": [t.get("tagFQN", "") for t in s.get("tags", [])],
            "owner": s.get("owner", {}).get("name", "Unknown") if s.get("owner") else "Unknown"
        })

    if tables_data:
        context = "REAL tables from OpenMetadata:\n"
        for t in tables_data:
            context += f"- {t['name']}: {t['description']}\n"
    else:
        context = "No tables found."

    response = groq_client.chat.completions.create(
        model="llama-3.1-8b-instant",
        messages=[
            {
                "role": "system",
                "content": "You are MetaChat, AI assistant for a data catalog. Use REAL table names. Be concise, max 3 sentences. Mention PII risks for customer/address tables."
            },
            {
                "role": "user",
                "content": f"{context}\n\nQuestion: {user_message}"
            }
        ],
        max_tokens=300
    )
    return {"reply": response.choices[0].message.content, "tables": tables_data}

if __name__ == "__main__":
    import uvicorn
    import webbrowser
    import threading
    import os

    # Auto open index.html after 2 seconds
    def open_browser():
        import time
        time.sleep(2)
        index_path = os.path.join(os.path.dirname(os.path.abspath(__file__)), "index.html")
        webbrowser.open(f"file:///{index_path}")
        print("🌐 Opened index.html in browser!")

    threading.Thread(target=open_browser, daemon=True).start()
    uvicorn.run(app, host="0.0.0.0", port=8000, log_level="warning")