requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.1
groq==0.4.2
python-dotenv==1.0.0
requests==2.31.0
